requirements.txt
gradio== 5.29.0
replicate== 1.0.6
Pillow== 11.2.1
python-dotenv== 1.1.0
requests== 2.32.3
pydantic== 2.10.6
pydantic_core== 2.27.2

# Testing dependencies
pytest== 8.2.0
pytest-cov== 6.1.1
pytest-mock== 3.14.0
coverage== 7.5.3